refactor: consolidate remaining lib files into services and features
