+++ frontend/webui/src/components/Config.tsx
@@ -6,8 +6,8 @@ import { useState, useEffect, useCallback, FormEvent } from 'react';
 import { useTranslation } from 'react-i18next';
 import { getSystemConfig, putSystemConfig, type AppConfig, type ConfigUpdate } from '../client-ts';
 import { notifyAuthRequiredIfUnauthorizedResponse } from '../lib/httpProblem';
-import { isUnauthorizedError } from '../lib/sessionEvents';
-import { unwrapClientResultOrThrow } from '../lib/clientWrapper';
+import { isUnauthorizedError } from '../features/player/sessionEvents';
+import { unwrapClientResultOrThrow } from '../services/clientWrapper';
 import { debugError, formatError } from '../utils/logging';
 import { getStoredToken } from '../utils/tokenStorage';
 import { Button, StatusChip } from './ui';
