+++ frontend/webui/src/components/Dashboard.test.tsx
@@ -1,7 +1,7 @@
 import type { ReactNode } from 'react';
 import { fireEvent, render, screen } from '@testing-library/react';
 import { describe, expect, it, beforeEach, vi } from 'vitest';
-import { ClientRequestError } from '../lib/clientWrapper';
+import { ClientRequestError } from '../services/clientWrapper';
 import Dashboard from './Dashboard';
 
 const mockNavigate = vi.fn();
