+++ frontend/webui/src/features/player/sessionEvents.ts
